hmi/requirements.txt
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dash==3.0.2
dash-bootstrap-components==2.0.1
dash-core-components==2.0.0
dash-enterprise-auth==0.2.5
dash-html-components==2.0.0
dash-table==5.0.0
dash_auth==2.3.0
Flask==3.0.3
Flask-Login==0.6.3
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
narwhals==1.30.0
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
plotly==6.0.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
retrying==1.3.4
setuptools==76.0.0
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.0.6
zipp==3.21.0
